page check function added
Added a function to check to see if a certain URL exists. LoL patch notes website doesn't use a traditional 404 error, so the function does a quick to check for the "patch-notes-container' html id which all patch notes pages have. If it doesn't find it, it doesn't attempt to parse the page.